[python] move to grid/ subdir and update
