--- python/dune/xt/bindings.cc
@@ -1,33 +1,0 @@
-// This file is part of the dune-xt-grid project:
-//   https://github.com/dune-community/dune-xt-grid
-// Copyright 2009-2018 dune-xt-grid developers and contributors. All rights reserved.
-// License: Dual licensed as BSD 2-Clause License (http://opensource.org/licenses/BSD-2-Clause)
-//      or  GPL-2.0+ (http://opensource.org/licenses/gpl-license)
-//          with "runtime exception" (http://www.dune-project.org/license.html)
-// Authors:
-//   Felix Schindler (2016 - 2017)
-//   René Fritze     (2018)
-
-#include "config.h"
-
-#include <string>
-#include <vector>
-
-#include <dune/common/parallel/mpihelper.hh>
-
-#include <dune/pybindxi/pybind11.h>
-
-#include <python/dune/xt/common/exceptions.bindings.hh>
-
-#include <python/dune/xt/common/bindings.hh>
-
-
-PYBIND11_MODULE(_grid, m)
-{
-  namespace py = pybind11;
-  using namespace pybind11::literals;
-
-  py::module::import("dune.xt.common");
-
-  Dune::XT::Common::bindings::add_initialization(m, "dune.xt.grid");
-}
