+++ python/dune/xt/grid/types.cc
@@ -19,8 +19,10 @@
 #include <dune/pybindxi/pybind11.h>
 #include <dune/pybindxi/stl.h>
 
+#include <python/dune/xt/common/bindings.hh>
 #include <python/dune/xt/grid/available_types.hh>
 #include <python/dune/xt/grid/grids.bindings.hh>
+
 
 template <class Tuple = Dune::XT::Grid::bindings::AvailableTypes>
 void addbind_for_Grid(pybind11::module& m, std::vector<std::string>& available_types)
@@ -41,6 +43,9 @@ PYBIND11_MODULE(_types, m)
   namespace py = pybind11;
 
   py::module::import("dune.xt.common");
+
+  Dune::XT::Common::bindings::add_initialization(m, "dune.xt.grid", "_types");
+
   std::vector<std::string> available_types;
   addbind_for_Grid(m, available_types);
   m.attr("available_types") = available_types;
